refactor: UIを表す箇所はhome.dartに分離させた
