+++ lib/view/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import 'home.dart';
 import 'make_members.dart';
 
 void main() => runApp(MyApp());
@@ -15,44 +16,5 @@ class MyApp extends StatelessWidget {
           '/make_members': (BuildContext context) =>
               MakeMembersPage(title: 'メンバー作成ページ'),
         });
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            RaisedButton(
-              child: Text('メンバー作成'),
-              onPressed: () {
-                Navigator.of(context).pushNamed('/make_members');
-              },
-            ),
-            RaisedButton(
-              child: Text('グループ作成'),
-              onPressed: () {
-                Navigator.of(context).pushNamed('/make_members');
-              },
-            ),
-          ],
-        )
-      )
-    );
   }
 }
